backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
redis>=5.0.3
celery>=5.3.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.79
python-magic>=0.4.27
aiofiles>=23.2.1
pytest>=8.1.1
pytest-asyncio>=0.23.6
httpx>=0.27.0
slowapi>=0.1.9
tenacity>=8.2.3
structlog>=24.1.0
python-multipart>=0.0.9
